# openapi.json
{
   "openapi": "3.0.3",
   "info": {
      "title": "Glazbeni instrumenti",
      "description": "Opis nekih glazbenih instrumenata.",
      "public": true,
      "version": "v3",
      "license": "MIT",
      "contact": {
         "name": "renato37",
         "url": "https://github.com/renato37/Otvoreno-racunarstvo-laboratorij",
         "email": "renato.ciz19@gmail.com"
      }
   },
   "host": "localhost:8000",
   "schemes": [
      "http"
   ],
   "basePath": "/",
   "consumes": [
      "application/json"
   ],
   "produces": [
      "application/json"
   ],
   "securityDefinitions": {
      "Basic": {
         "type": "basic"
      }
   },
   "security": [{
      "Basic": []
   }],
   "paths": {
...
      "/instrument/{id}/": {
         "get": {
            "operationId": "instrument_read",
            "description": "Prikaz instrumenta s tim ID.",
            "parameters": [{
               "id": "id"
            }],
            "responses": {
               "201": {
                  "description": "Prikazano.",
                  "content": {
                     "*/*": {
                        "schema": {
                           "type": "Object",
                           "items": {
                              "$ref": "#/glazbeni_instrumenti_dump.sql"
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Pogreška pri zahtjevu."
               },
               "404": {
                  "description": "Nije pronađeno."
               }
            },
            "tags": [
               "instrument"
            ]
         },
         "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
         }]
      },
      "/instrumenti/": {
         "get": {
            "operationId": "instrumenti_list",
            "description": "Popis svih instrumenata.",
            "parameters": [],
            "responses": {
               "201": {
                  "description": "Prikazano.",
                  "content": {
                     "*/*": {
                        "schema": {
                           "type": "Object",
                           "items": {
                              "$ref": "#/glazbeni_instrumenti_dump.sql"
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Pogreška pri zahtjevu."
               },
               "404": {
                  "description": "Nije pronađeno."
               }
            },
            "tags": [
               "instrumenti"
            ]
         },
         "parameters": []
      },
...
   },
...
}